backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0
chromadb==0.4.18
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
guardrails-ai==0.4.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4